# cmd/cmgrd/swagger.yaml
swagger: "2.0"
info:
  description: "This is a minimal REST API for `cmgr` that allows front-ends for a CTF to remotely manage challenges and their templating for events.  It is not intended to replace `cmgr` entirely as it does not currently expose the challenge update process."
  version: "0.2.0"
  title: "cmgrd"
  contact:
    email: "code@acictf.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
schemes:
- "http"
paths:
  /challenges:
    get:
      parameters:
        - name: "tags"
          in: "query"
          description: "Seach for available challenges limiting to given tags"
          required: false
          type: array
          items:
            type: string
      tags: [challenges]
      produces: ["application/json"]
      summary: "Lists all available challenges"
      responses:
        "500":
          description: "A database error occurred in `cmgr`"
        "200":
          description: "A list of all valid challenge identifiers"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ChallengeListElement"
...
  /schemas:
    get:
      tags: [schemas]
      produces: ["application/json"]
      summary: "Gets the list of all current schemas"
      responses:
        "404":
          description: "Invalid path string"
        "500":
          description: "A database error occurred in `cmgr` (currently includes bad build identifiers)"
        "200":
          description: "The list of schemas"
          schema:
            type: array
            items:
              type: string
    post:
      tags: [schemas]
      produces: ["application/json"]
      summary: "Creates a schema with the given definition"
      parameters:
        - in: "body"
          name: "body"
          description: "templating information for `cmgr`"
          required: true
          schema:
            $ref: "#/definitions/SchemaDefinition"
      responses:
        "400":
          description: "Value for 'schema_name' did not match the name in the schema definition"
        "404":
          description: "Invalid path string"
        "500":
          description: "A database error occurred in `cmgr` (currently includes bad build identifiers)."
        "201":
          description: "Indicates schema created successfully"
...
definitions:
  ChallengeListElement:
    type: "object"
    properties:
      id:
        type: string
      source_checksum:
        type: integer
        format: int64
      metadata_checksum:
        type: integer
        format: int64
      solve_script:
        type: boolean
      delivery_type:
        type: string
        enum: [service, artifact_only, flag_only]
        description: "Derived runtime shape of the challenge; non-service challenges never have instances."
...
  SchemaDefinition:
    type: object
    properties:
      name:
        type: string
      flag_format:
        type: string
      challenges:
        type: object
        additionalProperties:
          $ref: "#/definitions/SchemaElement"
  SchemaElement:
    type: object
    properties:
      seeds:
        type: array
        items:
          type: integer
          format: int32
      instance_count:
        type: integer
        format: int32
...